Implement 'Course' in 'Student' and retrieve with 'School'
This commit adds a new 'Course' model in the 'Student' class to represent the courses a student is enrolled in. The 'Course' has properties of 'name', 'teacher', and 'room'. A method to add courses to a student's list was added to the 'Student' class. Also, 'hashCode', 'equals' and 'toString' methods for 'Course' were generated. A method to retrieve all the courses of a student by getting the matriculation number was added to the 'School' class. This update allows tracking of student courses.