Introduced 'Sack' class with add, get, remove methods
Added a new 'Sack' class to manage an array of integer values. This class includes methods to add a new integer, get a number at a specific index, and remove numbers either at a particular index or from the end. This facilitates more efficient and flexible manipulation of integer arrays within the application.